refactor: resolved some conversations
