+++ model/src/main/java/com/demal/model/data/entity/tours/ITour.kt
@@ -5,17 +5,17 @@ import com.demal.model.data.entity.tours.network.ImageResponse
 
 interface ITour {
     val id: Int
-    val title: String?
+    val title: String
     val subTitle: String?
-    val place: String?
+    val place: String
     val price: Int
-    val country: String?
+    val country: String
     val duration: Int
     val description: String?
     val travelersCount: Int
     val startDate: String?
     val categoryId: Int
-    val createdAt: String?
+    val createdAt: String
     val category: CategoryResponse
     val images: List<ImageResponse>
 }
