+++ repository/src/main/java/com/demal/repository/data_sources/RemoteDataSource.kt
@@ -1,7 +1,7 @@
 package com.demal.repository.data_sources
 
 import com.demal.model.data.entity.AddToWishListEntity
-import com.demal.model.data.entity.tours.network.Tour
+import com.demal.model.data.entity.tours.Tour
 import com.demal.model.data.entity.user.LoginRequest
 import com.demal.model.data.entity.user.LoginResponse
 import com.demal.model.data.entity.user.User
